requirements.txt
numpy
pyusb
python-usbtmc
Pillow
h5py